+++ src/components/vm/VMToolbar.tsx
@@ -1,10 +1,10 @@
 import { Grid } from "@material-ui/core"
 import React from "react"
 import { VMProgramRunner } from "../../../jacdac-ts/src/vm/runner"
-import VMRoles from "./VMRoles"
+import BlockRoles from "../blockly/BlockRoles"
 import VMRunnerButtons from "./VMRunnerButtons"
 import VMStartSimulatorButton from "./VMStartSimulatorButton"
-import VMFileButtons from "./VMFileButtons"
+import BlockFileButtons from "../blockly/BlockFileButtons"
 
 export default function VMToolbar(props: {
     runner: VMProgramRunner
@@ -20,12 +20,12 @@ export default function VMToolbar(props: {
             alignItems="center"
             alignContent="center"
         >
-            <VMFileButtons />
+            <BlockFileButtons />
             <VMRunnerButtons runner={runner} run={run} cancel={cancel} />
             <Grid item>
                 <VMStartSimulatorButton />
             </Grid>
-            <VMRoles />
+            <BlockRoles />
         </Grid>
     )
 }
